requirement.txt
einops==0.2.0
flask==1.1.2
pillow==8.0.1
torch===1.7.0
torchvision
requests==2.24.0
opencv-python==4.4.0.46
gdown==3.12.2
matplotlib==3.3.2